pyproject.toml
[project]
name = "simple-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.32.0",
    "python-dotenv>=1.2.2",
]
